+++ react/utils/index.ts
@@ -141,3 +141,7 @@ export const getMatchRoute = ({
 
   return route ?? ''
 }
+
+export const transformUserRouteId = (id: string): string => {
+  return id.replace('custom#', 'custom::')
+}