+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Admin user can set an external URL to redirect user when selecting a specific binding.
+### Fixed
+- Redirect to the correct alternate hreflang for custom pages.
+
 ## [1.0.0] - 2021-03-03
 
 ### Changed